libcli/security: Make add_sid_to_array_unique use a uin32_t counter
Logical consequence of the previous commit


Autobuild-User: Michael Adam <obnox@samba.org>
Autobuild-Date: Mon Feb  7 19:24:19 CET 2011 on sn-devel-104
(cherry picked from commit 1d58fcc9259d929093d381f0bd6690345e7c8040)

+++ libcli/security/util_sid.c
@@ -310,7 +310,7 @@ NTSTATUS add_sid_to_array(TALLOC_CTX *mem_ctx, const struct dom_sid *sid,
 NTSTATUS add_sid_to_array_unique(TALLOC_CTX *mem_ctx, const struct dom_sid *sid,
 				 struct dom_sid **sids, uint32_t *num_sids)
 {
-	size_t i;
+	uint32_t i;
 
 	for (i=0; i<(*num_sids); i++) {
 		if (dom_sid_compare(sid, &(*sids)[i]) == 0)
